DRF/pyproject.toml
[project]
name = "drf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-xml>=2.0.0",
    "djangorestframework-yaml>=2.0.0",
    "httpie>=3.2.4",
    "pygments>=2.19.2",
]
